Example with username/password
I very often connect to MSSQL using username/password instead of integrated windows authentication. Usage of UID/PWD in the connection string is well documented in the provided link to ODBC connection string parameters. However, I believe an example connection string is useful to GDAL users.

Maybe the connection string should be in the section "Connecting to a database"